+++ src/avatarManager.ts
@@ -150,7 +150,7 @@ export class AvatarManager {
         {
           hostname: "api.github.com",
           path: "/repos/" + owner + "/" + repo + "/commits/" + avatarRequest.commits[commitIndex],
-          headers: { "User-Agent": "neo-git-graph" },
+          headers: { "User-Agent": "beads-git-graph" },
           agent: false,
           timeout: 15000
         },
@@ -219,7 +219,7 @@ export class AvatarManager {
         {
           hostname: "gitlab.com",
           path: "/api/v4/users?search=" + avatarRequest.email,
-          headers: { "User-Agent": "neo-git-graph", "Private-Token": "w87U_3gAxWWaPtFgCcus" }, // Token only has read access
+          headers: { "User-Agent": "beads-git-graph", "Private-Token": "w87U_3gAxWWaPtFgCcus" }, // Token only has read access
           agent: false,
           timeout: 15000
         },
@@ -290,7 +290,7 @@ export class AvatarManager {
           {
             hostname: imgUrl.hostname,
             path: imgUrl.path,
-            headers: { "User-Agent": "neo-git-graph" },
+            headers: { "User-Agent": "beads-git-graph" },
             agent: false,
             timeout: 15000
           },