+++ src/config.ts
@@ -14,7 +14,7 @@ class Config {
   private workspaceConfiguration: vscode.WorkspaceConfiguration;
 
   constructor() {
-    this.workspaceConfiguration = vscode.workspace.getConfiguration("neo-git-graph");
+    this.workspaceConfiguration = vscode.workspace.getConfiguration("beads-git-graph");
   }
 
   public autoCenterCommitDetailsView() {